refactor: 서비스단 패키지 정리
