+++ src/main/java/com/codingchosun/backend/service/post/PostService.java
@@ -1,4 +1,4 @@
-package com.codingchosun.backend.service;
+package com.codingchosun.backend.service.post;
 
 import com.codingchosun.backend.constants.DeleteConstants;
 import com.codingchosun.backend.domain.Hashtag;
