+++ src/main/java/com/codingchosun/backend/service/user/ProfileService.java
@@ -1,4 +1,4 @@
-package com.codingchosun.backend.service;
+package com.codingchosun.backend.service.user;
 
 
 import com.codingchosun.backend.domain.Hashtag;
